pyproject.toml
[project]
name = "3d-print-failure-detection"
version = "0.1.0"
description = ""
authors = [
    {name = "p3rc1va1",email = "35celik.baha@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "ultralytics (>=8.3.134,<9.0.0)",
    "websocket (>=0.2.1,<0.3.0)",
    "selenium (>=4.32.0,<5.0.0)"
]

[tool.poetry]
packages = [{include = "3d_print_failure_detection", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
